+++ src/main/java/com/example/bank_app/repository/AuditLogRepository.java
@@ -3,5 +3,5 @@ package com.example.bank_app.repository;
 import com.example.bank_app.model.AuditLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditLogRepository extends JpaRepository<AuditLog, Integer> {
+public interface AuditLogRepository extends JpaRepository<AuditLog, Long> {
 }
