+++ src/main/java/com/example/bank_app/repository/BankAccountRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface BankAccountRepository extends JpaRepository<BankAccount, Integer> {
-    List<BankAccount> findAllByUserIdAndIsActiveTrue(Integer userId);
+public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
+    List<BankAccount> findAllByUserIdAndIsActiveTrue(Long userId);
     boolean existsByAccountNumber(String accountNumber);
-    boolean existsByAccountNumberAndUserId(String accountNumber, Integer userId);
+    boolean existsByAccountNumberAndUserId(String accountNumber, Long userId);
     Optional<BankAccount> findByAccountNumberAndIsActiveTrue(String accountNumber);
 }
